Do the batch simulation

This is the main function for the batch simulation.

modified_simulation(sim_fun, var_list, default_list, bigmemory = TRUE)

batch_simulation(var_grid, sim_fun, default_list, bigmemory = TRUE)

Arguments

sim_fun
The simulation function. See sim_fun_test for an example.

var_list
An var_list object generated by fill_in_struct.

default_list
A list of default values for sim_fun.

bigmemory
Use hash_big.matrix-class to store large matrices?

var_grid
A var_grid object. See make_var_grid.

Value

A batch_simulation object, also a data frame. The first column, var, is a list of var_list that contains all the variables; the second to the second last columns are the values of the variables; the last column is the output of the simulation function.

Functions

modified_simulation: Modify a single simulation.

Examples

test <- new_var_set()
test <- test %>%
  add_var("par1", "var1", 1, 2, 0.1) %>%
  add_var("par2", "var2", 1, 2, 0.1)
test_grid <- make_var_grid(test)
test_result <- batch_simulation(test_grid, sim_fun_test,
  default_list = list(
    par1 = list(var1 = 0),
    par2 = list(var2 = 0, var3 = 0)
  )
)
test_result

#> Output(s) from 121 simulations.