Make 2D static landscape plot for a single simulation output

Make 2D static landscape plot for a single simulation output

make_2d_static(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5)

Arguments

output

A matrix of simulation output.

x

The name of the target variable.

adjust, from, to

Passed to density.

Umax

The maximum displayed value of potential.

Value

A 2d_static_landscape object that describes the landscape of the system, including the smooth distribution and the landscape plot.