Create and modify variable sets for batch simulation

This function was deprecated because we decided to shift to a more consistent terminology. Previous par is renamed as arg (argument) and previous var is renamed as ele (element). For creating an arg_set function, please use new_arg_set().

A variable set contains the descriptions of the relevant variables in a batch simulation. Use new_var_set to create a var_set object, and use add_var to add descriptions of variables.

new_var_set()

add_var(var_set, par_name, var_name, start, end, by)

Arguments

var_set

A var_set object.

par_name, var_name

The name of the parameter and variable in the simulation function

start, end, by

The data points where you want to test the variables. Passed to seq.

Value

A var_set object.

Functions

new_var_set(): Create a var_set.

add_var(): Add a variable to the var_set.

See also

Examples

test <- new_var_set()
#> Warning: `new_var_set()` was deprecated in simlandr 0.2.0.
#> Please use `new_arg_set()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
test <- test %>%
  add_var("par1", "var1", 1, 2, 0.1) %>%
  add_var("par2", "var2", 1, 2, 0.1)
#> Warning: `add_var()` was deprecated in simlandr 0.2.0.
#> Please use `add_arg_ele()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.