A simple non-gradient simulation function for testing

This is a toy stochastic non-gradient system which can have multistability in some conditions. Model specification:

Usage

sim_fun_nongrad(
  initial = list(x1 = 0, x2 = 0, a = 1),
  parameter = list(b = 1, k = 1, S = 0.5, n = 4, lambda = 0.01, sigmasq1 = 8, sigmasq2 =
    8, sigmasq3 = 2),
  constrain_a = TRUE,
  amin = -0.3,
  amax = 1.8,
  length = 1e+05,
  stepsize = 0.01,
  seed = NULL,
  progress = TRUE
)

Arguments

initial, parameter

Two sets of parameters. initial contains the initial value of x1, x2, and a; parameter contains b,k,S,n,lambda, which control the model dynamics, and sigmasq1,sigmasq2,sigmasq3, which are the squares of \(\sigma_1,\sigma_2,\sigma_3\) and controls the amplitude of noise.

constrain_a

Should the value of a be constrained? (TRUE by default).

amin, amax

If constrain_a, the minimum and maximum values of a.

length

The length of simulation.

stepsize

The step size used in the Euler method.

seed

The initial seed that will be passed to set.seed() function.

progress

Show progress bar of the simulation?

Value

A matrix of simulation results.

Details

$$\frac {dx_ {1}}{dt} = \frac {ax_ {1}^ {n}}{S^ {n}+x_ {1}^ {n}} + \frac {bS^ {n}}{S^ {n}+x_ {2}^ {n}} - kx_ {1}+ \sigma_1 dW_1/dt$$ $$\frac {dx_ {2}}{dt} = \frac {ax_ {2}^ {n}}{S^ {n}+x_ {2}^ {n}} + \frac {bS^ {n}}{S^ {n}+x_ {1}^ {n}} - kx_ {2}+ \sigma_2 dW_2/dt$$ $$\frac {da}{dt} = -\lambda a+ \sigma_3 dW_3/dt$$

References

Wang, J., Zhang, K., Xu, L., & Wang, E. (2011). Quantifying the Waddington landscape and biological paths for development and differentiation. Proceedings of the National Academy of Sciences, 108(20), 8257-8262. doi:10.1073/pnas.1017017108

See also